Refactor finance module structure for cleaner imports
Updated directory structure and adjusted imports across the finance module to improve modularity and maintainability. Consolidated and relocated files to align with the new structure for better organization.
