+++ apps/nest-js/src/finance/bank/bank.controller.ts
@@ -4,10 +4,10 @@ import { AuthGuard } from '@nestjs/passport';
 import { ERole } from '@repo/business/enum';
 import { QueryParameters } from '@repo/business/types';
 
-import { AuthRoles } from '../../../decorators/auth-role/auth-roles.decorator';
+import { AuthRoles } from '../../decorators/auth-role/auth-roles.decorator';
 
-import { AuthRoleGuard } from '../../../guards/auth-role/auth-role.guard';
-import { AuthStatusGuard } from '../../../guards/auth-status/auth-status.guard';
+import { AuthRoleGuard } from '../../guards/auth-role/auth-role.guard';
+import { AuthStatusGuard } from '../../guards/auth-status/auth-status.guard';
 
 import { BankService } from './bank.service';
 import { CreateBankDto } from './dto/create-bank.dto';
@@ -15,7 +15,7 @@ import { UpdateBankDto } from './dto/update-bank.dto';
 
 
 
-@Controller('finance/bill')
+@Controller('finance')
 @UseGuards(AuthGuard(), AuthRoleGuard, AuthStatusGuard)
 export class BankController {
   constructor(private readonly service: BankService) {}
