+++ apps/nest-js/src/finance/bank/bank.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Bank } from '../../entities/bank.entity';
+import { Bank } from '../entities/bank.entity';
 import { BankController } from './bank.controller';
 import { BankService } from './bank.service';
