+++ apps/nest-js/src/finance/bank/dto/create-bank.dto.ts
