+++ apps/nest-js/src/finance/bill/expense/expense.module.ts
@@ -6,7 +6,7 @@ import ExpenseBusiness from '@repo/business/finance/expense/business/business';
 
 import { Expense } from '../../entities/expense.entity';
 import { ExpenseService } from './expense.service';
-import { SupplierModule } from './supplier/supplier.module';
+import { SupplierModule } from '../../supplier/supplier.module';
 
 @Module({
   imports: [
