+++ apps/nest-js/src/finance/category/category.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { BillCategory } from '../../entities/category.entity';
+import { BillCategory } from '../entities/category.entity';
 import { CategoryController } from './category.controller';
 import { CategoryService } from './category.service';
