+++ apps/nest-js/src/finance/category/category.service.spec.ts
@@ -4,9 +4,9 @@ import { ConflictException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { getRepositoryToken } from '@nestjs/typeorm';
 
-import { BILL_CATEGORY_MOCK } from '../../../mocks/bill-category.mock';
-import { BILL_MOCK } from '../../../mocks/bill.mock';
-import { BillCategory } from '../../entities/category.entity';
+import { BILL_CATEGORY_MOCK } from '../../mocks/bill-category.mock';
+import { BILL_MOCK } from '../../mocks/bill.mock';
+import { BillCategory } from '../entities/category.entity';
 
 import { CategoryService } from './category.service';
 import { type CreateCategoryDto } from './dto/create-category.dto';
