+++ apps/nest-js/src/finance/finance.module.ts
@@ -2,9 +2,12 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { BankModule } from './bank/bank.module';
 import { BillModule } from './bill/bill.module';
+import { CategoryModule } from './category/category.module';
 import { FinanceController } from './finance.controller';
 import { FinanceService } from './finance.service';
+import { SupplierModule } from './supplier/supplier.module';
 
 import { Finance } from './entities/finance.entity';
 
@@ -14,7 +17,10 @@ import { Finance } from './entities/finance.entity';
     imports: [
         TypeOrmModule.forFeature([Finance]),
         PassportModule.register({ defaultStrategy: 'jwt' }),
+        BankModule,
         BillModule,
+        CategoryModule,
+        SupplierModule
     ],
     exports: [FinanceService]
 })
