+++ apps/nest-js/src/finance/supplier/supplier.controller.spec.ts
@@ -1,8 +1,8 @@
 import { Test, type TestingModule } from '@nestjs/testing';
 import {afterEach, beforeEach, describe, expect, it, jest} from '@jest/globals';
 
-import { SUPPLIER_MOCK } from '../../../../mocks/supplier.mock';
-import { type Supplier } from '../../../entities/supplier.entity';
+import { SUPPLIER_MOCK } from '../../mocks/supplier.mock';
+import { type Supplier } from '../entities/supplier.entity';
 
 import { type CreateSupplierDto } from './dto/create-supplier.dto';
 import { SupplierController } from './supplier.controller';
