+++ apps/nest-js/src/finance/supplier/supplier.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Supplier } from '../../../entities/supplier.entity';
+import { Supplier } from '../entities/supplier.entity';
 import { SupplierController } from './supplier.controller';
 import { SupplierService } from './supplier.service';
 import { TypeModule } from './type/type.module';
