+++ apps/nest-js/src/finance/supplier/supplier.service.ts
@@ -6,12 +6,12 @@ import { transformObjectDateAndNulls } from '@repo/services/object/object';
 
 import SupplierConstructor from '@repo/business/finance/supplier/supplier';
 
-import { Service } from '../../../../shared';
+import { Service } from '../../shared';
 
-import type { FinanceSeederParams } from '../../../types';
+import type { FinanceSeederParams } from '../types';
 
-import { Supplier } from '../../../entities/supplier.entity';
-import { SupplierType } from '../../../entities/type.entity';
+import { Supplier } from '../entities/supplier.entity';
+import { SupplierType } from '../entities/type.entity';
 
 import { CreateSupplierDto } from './dto/create-supplier.dto';
 import { SupplierTypeService } from './type/type.service';
