+++ apps/nest-js/src/finance/supplier/type/type.service.spec.ts
@@ -4,9 +4,9 @@ import { ConflictException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { getRepositoryToken } from '@nestjs/typeorm';
 
-import { SUPPLIER_MOCK } from '../../../../../mocks/supplier.mock';
-import { SUPPLIER_TYPE_MOCK } from '../../../../../mocks/supplier-type.mock';
-import { SupplierType } from '../../../../entities/type.entity';
+import { SUPPLIER_MOCK } from '../../../mocks/supplier.mock';
+import { SUPPLIER_TYPE_MOCK } from '../../../mocks/supplier-type.mock';
+import { SupplierType } from '../../entities/type.entity';
 
 import { type CreateTypeDto } from './dto/create-type.dto';
 import { SupplierTypeService } from './type.service';
